requirements.txt
flask==3.0.3
gunicorn==22.0.0
anthropic>=0.40.0
gspread==6.1.2
google-auth==2.30.0
requests==2.32.3
beautifulsoup4==4.12.3
apscheduler==3.10.4
